apps/worker/python/requirements.txt
audio-separator[gpu]==0.28.5
onnxruntime-openvino==1.19.0
openvino==2024.4.0
yt-dlp
